--- a/pyproject.toml
+++ b/pyproject.toml
@@ -21,6 +21,7 @@ Homepage = "https://github.com/GarryBGoode/gggears"
 numpy = "^2.0.0"
 scipy = "^1.10.1"
 build123d = "^0.8.0"
+joblib = "^1.4.2"
 
 [project.optional-dependencies]
 dev = [