--- a/requirements.txt
+++ b/requirements.txt
@@ -6,3 +6,4 @@ pytest-xdist
 shapely
 matplotlib
 ocp-vscode
+joblib