parallelized GearToNurb::generate_nurbs(), GearToNurb::generate_gear_stacks() and GearToNurb::generate_surface_points_sides()

The code base also compiles with dev branch of build123d as at 20 January 2025.

fixes #4 

TODO: use perf profiler to identify other performance bottlenecks

https://docs.python.org/3.12/howto/perf_profiling.html

@GarryBGoode Are you aware of any other computation-intensive for-loops, or scipy / numpy calls in anywhere of the gggears code?  I am only new to gggears, so I am not yet familiar with its algorithms.