@GarryBGoode 2025-01-19
@kcleung Did you do any benchmarking that it got any faster?