@GarryBGoode 2025-01-19
I'm seeing mixed results, a little biased towards worse performance on my computer :(
Though I'm not sure if I have something configured wrong.

Example `bevel chain`
baseline
```2025-01-19 21:08:05,020 - INFO - Gears generated in 0.29088 seconds
2025-01-19 21:08:05,236 - INFO - Nurbs generated in 0.21604 seconds
2025-01-19 21:08:05,242 - INFO - Surfaces generated in 0.00554 seconds```
joblibbed
```
2025-01-19 21:09:31,985 - INFO - Gears generated in 2.50459 seconds
2025-01-19 21:09:33,905 - INFO - Nurbs generated in 1.91921 seconds
2025-01-19 21:09:35,711 - INFO - Surfaces generated in 1.80614 seconds
```

Admittedly, this example uses a low number of z-vals and z-tweens.


Example `fishbone_bevels`
baseline
```
2025-01-19 21:14:44,563 - INFO - Gears generated in 2.23833 seconds
2025-01-19 21:14:45,950 - INFO - Nurbs generated in 1.38682 seconds
2025-01-19 21:14:47,376 - INFO - Surfaces generated in 1.42684 seconds
```
joblibbed
```
2025-01-19 21:13:23,974 - INFO - Gears generated in 3.61325 seconds
2025-01-19 21:13:26,589 - INFO - Nurbs generated in 2.61496 seconds
2025-01-19 21:13:29,331 - INFO - Surfaces generated in 2.74185 seconds
```

Example `worm_approx`
baseline
```
2025-01-19 21:17:43,068 - INFO - Gears generated in 3.22286 seconds
2025-01-19 21:17:44,626 - INFO - Nurbs generated in 1.55718 seconds
2025-01-19 21:18:10,598 - INFO - Surfaces generated in 25.97159 seconds
2025-01-19 21:18:10,949 - INFO - Gears generated in 0.16604 seconds
2025-01-19 21:18:11,245 - INFO - Nurbs generated in 0.29607 seconds
2025-01-19 21:18:11,284 - INFO - Surfaces generated in 0.03901 seconds
```

joblibbed
```
2025-01-19 21:16:33,502 - INFO - Gears generated in 3.76423 seconds
2025-01-19 21:16:36,326 - INFO - Nurbs generated in 2.82433 seconds
2025-01-19 21:16:51,726 - INFO - Surfaces generated in 15.39943 seconds
2025-01-19 21:16:53,846 - INFO - Gears generated in 1.93263 seconds
2025-01-19 21:16:55,947 - INFO - Nurbs generated in 2.10043 seconds
2025-01-19 21:16:57,763 - INFO - Surfaces generated in 1.81638 seconds
```
This last one has n_z_tweens = 29 and 3 stacks, so it needs to generate 29*3=87 gear profiles and nurb profiles, and then run the minimize for the surface data 3 times. Still, it seems only the surface generation benefited from parallelization.